pyproject.toml
[project]
name = "deepagents"
version = "0.2.5"
description = "General purpose 'deep agent' with sub-agent spawning, todo list capabilities, and mock file system. Built on LangGraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain-anthropic>=1.0.0,<2.0.0",
    "langchain>=1.0.2,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-openai>=0.1.0",
    "wcmatch",
    "tavily>=1.1.0",
    "research-agent>=0.0.2",
    "clibuilder>=0.1.7",
    "agents>=1.4.0",
    "pyfiglet>=1.0.4",
    "httpx",
    "pdfplumber>=0.11.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv",
]


[project.scripts]
hkex = "src.cli.main:cli_main"

[project.urls]
Homepage = "https://docs.langchain.com/oss/python/deepagents/overview"
Documentation = "https://reference.langchain.com/python/deepagents/"
Source = "https://github.com/langchain-ai/deepagents"
Twitter = "https://x.com/LangChainAI"
Slack = "https://www.langchain.com/join-community"
Reddit = "https://www.reddit.com/r/LangChain/"


[dependency-groups]
test = [
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "ruff>=0.12.2,<0.13.0",
  "mypy>=1.18.1,<1.19.0",
]

dev = [
  "langchain-openai",
  "twine",
  "build",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
[tool.setuptools.packages.find]
where = ["libs", "."]
include = ["deepagents*", "src", "src.*"]
exclude = ["deepagents.tests*", "deepagents.tests.*", "tests*", "tests.*", "*.egg-info*"]

[tool.setuptools.package-data]
"*" = ["py.typed", "*.md"]

[tool.ruff]
line-length = 150
# Exclude any files that shouldn't be linted
exclude = []

[tool.ruff.format]
docstring-code-format = true  # Formats code blocks in docstrings

[tool.ruff.lint]
select = [
    "ALL"  # Enable all rules by default
]
ignore = [
    "COM812",  # Messes with the formatter
    "ISC001",  # Messes with the formatter
    "PERF203", # Rarely useful
    "SLF001",  # Private member access
    "PLC0415", # Imports should be at the top. Not always desirable
    "PLR0913", # Too many arguments in function definition
    "PLC0414", # Inconsistent with how type checkers expect to be notified of intentional re-exports
    "C901",    # Too complex
]
unfixable = ["B028"]  # Rules that shouldn't be auto-fixed

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"  # Google-style docstrings
ignore-var-parameters = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D1",      # Skip documentation rules in tests
    "S101",    # Allow asserts in tests
    "S311",    # Allow pseudo-random generators in tests
    # Add more test-specific ignores
]

[tool.mypy]
strict = true
ignore_missing_imports = true
enable_error_code = ["deprecated"]
# Optional: reduce strictness if needed
disallow_any_generics = false
warn_return_any = false

[tool.uv.workspace]
members = [
    "libs/deepagents-cli",
]
